feat(tcp): option negotiation, window scale, timestamps/PAWS
握手で MSS/window scale/timestamps/SACK-permitted を折衝し TCB に保存。
window scale は両側が送った時のみ有効・shift>14 は 14 に clamp・SYN の window は
生値。受信窓 65535 超 (最大 2^30) を扱える。timestamps で TSecr から RTT を測り
(Karn の例外) PAWS で古いセグメントを弾く。TS.Recent は wrap 考慮で単調更新。

+++ tcp/statemachine_test.go
@@ -179,7 +179,7 @@ func estab(t *testing.T) (*Conn, Link, *fakeClock) {
 	c.tcb.rcv.nxt = 8000
 	c.tcb.rcv.wnd = 1000
 	c.tcb.rcv.irs = 7999
-	c.tcb.maxSndWnd = maxWindow
+	c.tcb.maxSndWnd = uint32(maxWindow)
 	return c, peer, fc
 }
